@@ -2,12 +2,8 @@ package com.example.buensabor.security.controllers;
 
 import com.example.buensabor.dtos.Message;
 import com.example.buensabor.security.dto.*;
-import com.example.buensabor.security.entities.Rol;
 import com.example.buensabor.security.entities.Usuario;
-import com.example.buensabor.security.enums.RolNombre;
-import com.example.buensabor.security.jwt.JwtProvider;
 import com.example.buensabor.security.services.EnviarMailService;
-import com.example.buensabor.security.services.RolService;
 import com.example.buensabor.security.services.UsuarioService;
 import com.google.api.client.googleapis.auth.oauth2.GoogleIdToken;
 import com.google.api.client.googleapis.auth.oauth2.GoogleIdTokenVerifier;
@@ -15,15 +11,10 @@ import com.google.api.client.http.javanet.NetHttpTransport;
 import com.google.api.client.json.jackson2.JacksonFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
-import javax.mail.MessagingException;
-import javax.validation.Valid;
+import jakarta.mail.MessagingException;
+import jakarta.validation.Valid;
 import java.io.IOException;
 import java.text.ParseException;
 import java.util.*;
@@ -35,11 +26,7 @@ import org.springframework.http.ResponseEntity;
 @RequestMapping("/auth")
 public class UsuarioController {
 
-    private final PasswordEncoder passwordEncoder;
-    private final AuthenticationManager authenticationManager;
     private final UsuarioService usuarioService;
-    private final RolService rolService;
-    private final JwtProvider jwtProvider;
     private final EnviarMailService emailSenderService;
     @Value("${google.clientId}")
     private String googleClientId;
@@ -48,23 +35,19 @@ public class UsuarioController {
     @Value("${spring.mail.username}")
     private String mailFrom;
     private static final String SUBJECT = "Bienvenido, solo queda verificar tu correo.";
-    @Value("${mail.urlVerification}")
+    @Value("${mail.url.verification}")
     private String urlVerification;
 
     @Autowired
-    public UsuarioController(PasswordEncoder passwordEncoder, AuthenticationManager authenticationManager, UsuarioService usuarioService, RolService rolService, JwtProvider jwtProvider, EnviarMailService emailSenderService) {
-        this.passwordEncoder = passwordEncoder;
-        this.authenticationManager = authenticationManager;
+    public UsuarioController(UsuarioService usuarioService, EnviarMailService emailSenderService) {
         this.usuarioService = usuarioService;
-        this.rolService = rolService;
-        this.jwtProvider = jwtProvider;
         this.emailSenderService = emailSenderService;
     }
 
     @PostMapping("/update")
     public Usuario update(@RequestBody Usuario usuario){
         usuario.setEnabled(true);
-        return usuarioService.save(usuario);
+        return usuarioService.update(usuario);
     }
 
     @PostMapping("/google")
@@ -77,38 +60,39 @@ public class UsuarioController {
         final GoogleIdToken.Payload payload = googleIdToken.getPayload();
         Usuario usuario;
         if (usuarioService.existsByEmail(payload.getEmail())) {
-            usuario = usuarioService.getByEmail(payload.getEmail()).get();
+            usuario = usuarioService.getByEmail(payload.getEmail()).orElseThrow();
         } else {
-            NuevoUsuario nuevoUsuario = new NuevoUsuario();
-            nuevoUsuario.setNombre((String) payload.get("given_name"));
-            nuevoUsuario.setApellido((String) payload.get("family_name"));
-            nuevoUsuario.setTelefono("");
-            nuevoUsuario.setEmail(payload.getEmail());
-            nuevoUsuario.setClave(secretPsw);
-            nuevoUsuario.setEnabled(true);
-            usuario = saveNewUsuario(nuevoUsuario);
+            NuevoUsuarioDto nuevoUsuarioDto = new NuevoUsuarioDto();
+            nuevoUsuarioDto.setNombre((String) payload.get("given_name"));
+            nuevoUsuarioDto.setApellido((String) payload.get("family_name"));
+            nuevoUsuarioDto.setTelefono("");
+            nuevoUsuarioDto.setEmail(payload.getEmail());
+            nuevoUsuarioDto.setClave(secretPsw);
+            nuevoUsuarioDto.setEnabled(true);
+            usuario = usuarioService.save(nuevoUsuarioDto);
         }
-        JwtDto jwtDto = loginUsuario(usuario.getEmail(), secretPsw);
+        LoginUsuarioDto loginUsuarioDto = new LoginUsuarioDto(usuario.getEmail(), secretPsw);
+        JwtDto jwtDto = usuarioService.login(loginUsuarioDto);
         return new ResponseEntity<>(jwtDto, HttpStatus.OK);
     }
 
     @PostMapping("/nuevo")
-    public ResponseEntity<?> nuevo(@Valid @RequestBody NuevoUsuario nuevoUsuario, BindingResult bindingResult) throws MessagingException {
+    public ResponseEntity<?> nuevo(@Valid @RequestBody NuevoUsuarioDto nuevoUsuarioDto, BindingResult bindingResult) throws MessagingException {
         if (bindingResult.hasErrors()) {
             return new ResponseEntity<>(new Message("campos mal puestos o email inválido"), HttpStatus.BAD_REQUEST);
         }
-        if (usuarioService.existsByEmail(nuevoUsuario.getEmail())) {
+        if (usuarioService.existsByEmail(nuevoUsuarioDto.getEmail())) {
             return new ResponseEntity<>(new Message("ese email ya existe"), HttpStatus.BAD_REQUEST);
         }
         UUID uuid = UUID.randomUUID();
         String tokenPassword = uuid.toString();
-        nuevoUsuario.setEnabled(false);
-        nuevoUsuario.setTokenPassword(tokenPassword);
-        saveNewUsuario(nuevoUsuario);
+        nuevoUsuarioDto.setEnabled(false);
+        nuevoUsuarioDto.setTokenPassword(tokenPassword);
+        usuarioService.save(nuevoUsuarioDto);
         EmailValuesDto emailValuesDto = new EmailValuesDto();
-        emailValuesDto.setNombre(nuevoUsuario.getNombre());
+        emailValuesDto.setNombre(nuevoUsuarioDto.getNombre());
         emailValuesDto.setMailFrom(mailFrom);
-        emailValuesDto.setMailTo(nuevoUsuario.getEmail());
+        emailValuesDto.setMailTo(nuevoUsuarioDto.getEmail());
         emailValuesDto.setSubject(SUBJECT);
 
         emailValuesDto.setTokenPassword(tokenPassword);
@@ -117,22 +101,20 @@ public class UsuarioController {
     }
 
     @PostMapping("/login")
-    public ResponseEntity<JwtDto> login(@Valid @RequestBody LoginUsuario loginUsuario, BindingResult bindingResult) {
-        Usuario usuario = usuarioService.getByEmail(loginUsuario.getEmail()).get();
+    public ResponseEntity<?> login(@Valid @RequestBody LoginUsuarioDto loginUsuarioDto, BindingResult bindingResult) {
+        Usuario usuario = usuarioService.getByEmail(loginUsuarioDto.getEmail()).orElseThrow();
         if (bindingResult.hasErrors()) {
-            return new ResponseEntity(new Message("Campos mal puestos o email inválido"), HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(new Message("Campos mal puestos o email inválido"), HttpStatus.BAD_REQUEST);
         }
         if (!usuario.isEnabled()){
-            return new ResponseEntity(new Message("Cuenta no verificada"), HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(new Message("Cuenta no verificada"), HttpStatus.BAD_REQUEST);
         }
-        return new ResponseEntity<>(loginUsuario(loginUsuario.getEmail(), loginUsuario.getClave()), HttpStatus.OK);
+        return new ResponseEntity<>(usuarioService.login(loginUsuarioDto), HttpStatus.OK);
     }
 
     @PostMapping("/refresh")
-    public ResponseEntity<JwtDto> refreshToken(@RequestBody JwtDto jwtDto) throws ParseException {
-        String token = jwtProvider.refreshToken(jwtDto);
-        JwtDto jwt = new JwtDto(token);
-        return new ResponseEntity<>(jwt, HttpStatus.OK);
+    public ResponseEntity<?> refreshToken(@RequestBody JwtDto jwtDto) throws ParseException {
+        return new ResponseEntity<>(usuarioService.refreshToken(jwtDto), HttpStatus.OK);
     }
 
     @GetMapping("/{email}")
@@ -140,36 +122,4 @@ public class UsuarioController {
         return usuarioService.getByEmail(email);
     }
 
-
-
-    private JwtDto loginUsuario(String email, String password) {
-        Authentication authentication = authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(email, password));
-        SecurityContextHolder.getContext().setAuthentication(authentication);
-        String jwt = jwtProvider.generarToken(authentication);
-        return new JwtDto(jwt);
-    }
-
-    private Usuario saveNewUsuario(NuevoUsuario nuevoUsuario) {
-        Usuario usuario = new Usuario(
-                nuevoUsuario.getNombre(),
-                nuevoUsuario.getApellido(),
-                nuevoUsuario.getTelefono(),
-                nuevoUsuario.getEmail(),
-                passwordEncoder.encode(nuevoUsuario.getClave()),
-                nuevoUsuario.isEnabled(),
-                nuevoUsuario.getTokenPassword());
-        Set<Rol> roles = new HashSet<>();
-        roles.add(rolService.getByRolNombre(RolNombre.ROLE_USER).get());
-        if (nuevoUsuario.getRoles().contains("admin")) {
-            roles.add(rolService.getByRolNombre(RolNombre.ROLE_ADMIN).get());
-        }
-        if (nuevoUsuario.getRoles().contains("cocinero")) {
-            roles.add(rolService.getByRolNombre(RolNombre.ROLE_COCINERO).get());
-        }
-        if (nuevoUsuario.getRoles().contains("cajero")) {
-            roles.add(rolService.getByRolNombre(RolNombre.ROLE_CAJERO).get());
-        }
-        usuario.setRoles(roles);
-        return usuarioService.save(usuario);
-    }
 }
