@@ -5,8 +5,8 @@ import com.example.buensabor.dtos.RubroDto;
 import com.example.buensabor.entities.rubro.Rubro;
 import com.example.buensabor.services.rubro.RubroService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/v1/rubros")
@@ -18,13 +18,13 @@ public class RubroController extends BaseController<Rubro, RubroDto, RubroServic
     }
 
     @GetMapping("/todosActivos")
-    public List<Rubro> getRubrosSinFechaDeBaja() {
-        return service.getRubrosSinFechaDeBaja() ;
+    public ResponseEntity<?> getRubrosSinFechaDeBaja() {
+        return ResponseEntity.ok(service.getRubrosSinFechaDeBaja());
     }
 
 
     @GetMapping("/rubro")
-    public List<Rubro> getRubroByIdNotInsumo(@RequestParam("id") Long id) {
-        return service.getRubroByIdArticuloInsumo(id);
+    public ResponseEntity<?> getRubroByIdNotInsumo(@RequestParam("id") Long id) {
+        return ResponseEntity.ok(service.getRubroByIdArticuloInsumo(id));
     }
 }
