@@ -1,6 +1,6 @@
 package com.example.buensabor.factories;
 
-public interface IFactory <E, DTO>{
+public interface Factory<E, DTO>{
     E createEntity (DTO dto);
     DTO createDto (E entity);
 }
