@@ -6,8 +6,9 @@ import org.springframework.context.annotation.Configuration;
 
 @Configuration
 public class ModelMapperConfig {
+
     @Bean
-    public ModelMapper getModelMapper(){
+    public ModelMapper modelMapper(){
         return new ModelMapper();
     }
 }
