@@ -1,5 +1,5 @@
 package com.example.buensabor.security.enums;
 
 public enum RolNombre {
-    ROLE_ADMIN, ROLE_USER, ROLE_COCINERO, ROLE_CAJERO;
+    ROLE_ADMIN, ROLE_USER
 }
